+++ algorithm/src/main/kotlin/io/github/wliamp/agr/config/AlgorithmAutoConfig.kt
@@ -1,6 +1,6 @@
 package io.github.wliamp.agr.config
 
-import io.github.wliamp.agr.util.MatchMaker
+import io.github.wliamp.agr.MatchMaker
 import org.springframework.boot.autoconfigure.AutoConfiguration
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean
 import org.springframework.context.annotation.Bean
