+++ algorithm/src/main/kotlin/io/github/wliamp/agr/data/Target.kt
@@ -1,5 +1,7 @@
 package io.github.wliamp.agr.data
 
+import io.github.wliamp.agr.impl.ICriteria
+
 data class Target(
     val id: String,
     val criteria: ICriteria
