+++ algorithm/src/main/kotlin/io/github/wliamp/agr/impl/ICriteria.kt
@@ -1,4 +1,4 @@
-package io.github.wliamp.agr.data
+package io.github.wliamp.agr.impl
 
 interface ICriteria {
     fun matches(other: ICriteria): Boolean
